toystory_AI/requirements.txt
accelerate==0.32.1
aiofiles==23.2.1
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
aspose-3d==24.7.0
async-timeout==4.0.3
attrs==24.2.0
boto3==1.35.3
botocore==1.35.3
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
datasets==2.14.7
dill==0.3.7
distro==1.9.0
einops==0.7.0
exceptiongroup==1.2.2
fastapi==0.112.1
ffmpy==0.4.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2023.10.0
glcontext==2.5.0
GPUtil==1.4.0
gradio==4.41.0
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.4
imageio==2.35.0
imageio-ffmpeg==0.5.1
importlib_resources==6.4.2
Jinja2==3.1.3
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
moderngl==5.10.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
narwhals==1.4.2
networkx==3.2.1
numba==0.60.0
numpy==1.26.3
objaverse==0.1.7
omegaconf==2.3.0
onnxruntime==1.19.0
openai==1.42.0
opencv-python-headless==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.2
Pillow==10.1.0
pipdeptree==2.23.1
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.3
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PyMatting==1.1.12
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
rembg==2.0.58
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
ruff==0.6.1
s3transfer==0.10.2
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
sympy==1.12
threadpoolctl==3.5.0
tifffile==2024.8.10
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchmcubes @ git+https://github.com/tatsy/torchmcubes.git@cb81cddece46a8a126b08f7fbb9742f8605eefab
torchvision==0.16.0+cu121
tqdm==4.66.5
transformers==4.44.0
trimesh==4.0.5
triton==2.1.0
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
uvicorn==0.30.6
websockets==11.0.3
xatlas==0.0.9
xxhash==3.5.0
yarl==1.9.4
zipp==3.20.0
